pyproject.toml
[project]
name = "pptx-accessibility-checker"
version = "0.1.0"
description = "PowerPoint & PDF accessibility automation application"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",

    # PPTX Processing
    "python-pptx>=1.0.2",
    "Pillow>=11.0.0",

    # PDF Processing
    "pikepdf>=9.4.2",
    "pypdfium2>=4.30.1",
    "pdfplumber>=0.11.4",
    "pytesseract>=0.3.13",

    # AI & Vision
    "anthropic>=0.42.0",

    # Accessibility Analysis
    "colormath>=3.0.0",
    "webcolors>=24.11.1",

    # Report Generation
    "weasyprint>=63.1",
    "jinja2>=3.1.4",
    "markdown>=3.7",

    # Core
    "pydantic>=2.10.5",
    "pydantic-settings>=2.7.0",
    "aiofiles>=24.1.0",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.0.1",
    "loguru>=0.7.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "httpx>=0.28.1",
    "black>=24.10.0",
    "ruff>=0.8.4",
    "mypy>=1.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pptx_accessibility"]

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
